# proto/example.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "example.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/example/echo": {
      "post": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoStringMessage"
            }
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    },
    "/v1/messages/{message_id}": {
      "get": {
        "operationId": "GetMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "sub.subfield",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    }
  },
  "definitions": {
    "GetMessageRequestSubMessage": {
      "type": "object",
      "properties": {
        "subfield": {
          "type": "string"
        }
      }
    },
    "protoMessage": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      }
    },
    "protoStringMessage": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    }
  }
}